[FRAMEWORK] Updates to allow OpenVX to do IPC between 32b and 64b systems
 Changes shared pointers from uintptr_t and void* to uint64_t

 Tested in x86_64 HOST emulation mode and on 32b+64b sim environment.
 NOT tested on TDA2x and x86 32b HOST emulation mode

+++ source/framework/vx_obj_desc.c
@@ -141,7 +141,7 @@ tivx_obj_desc_t *tivxObjDescAlloc(vx_enum type, vx_reference ref)
             /* init entry that is found */
             tmp_obj_desc->obj_desc_id = idx;
             tmp_obj_desc->type = type;
-            tmp_obj_desc->host_ref = (uintptr_t)ref;
+            tmp_obj_desc->host_ref = (uint64_t)ref;
 
             g_obj_desc_table.last_alloc_index
                 = (idx+1)%g_obj_desc_table.num_entries;
